+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - cli: Add `todo reschedule` command
+- cli: Set fields in `new`, `edit` command, and skip TUI if all required fields provided
 - core: Add extra `DateTime` and `Time` for `DateTimeAnchor`
 - core: Convert `DateTimeAnchor` to `LooseDateTime`
 - core: Parse as hours (e.g., "10h", "10 hours", "in 10hours"), days (e.g., "10d", "in 10d",